More disassembling. Fixed format 3

+++ thumb/src/asm/data.rs
@@ -106,7 +106,7 @@ impl Display for Register {
 #[derive(Debug)]
 pub enum Operand {
     Register(Register),
-    Immediate(u8),
+    Immediate(u32),
 }
 
 #[derive(Debug)]
